requirements.txt
aiohttp
aiosqlite
discord.py==2.6.3
python-dotenv
hikari
google-generativeai
textual
